Improve specs to be more readable
Effort is made to follow the best practices in http://betterspecs.org/

+++ spec/database_spec.rb
@@ -61,50 +61,48 @@ describe Viva::Database do
   end
 
   describe '#add' do
-    context 'track only' do
-      it 'adds one track to the database at a time' do
+    context 'when adding tracks' do
+      it 'adds one' do
         db.add({track: tracks[0]})
-        db.add({track: tracks[0]})
-        expect(Viva::Database::Track.count).to eq 2
+        expect(Viva::Database::Track.count).to eq 1
       end
 
-      it 'adds multiple tracks to the database' do
+      it 'adds multiple' do
         db.add({track: tracks})
         expect(Viva::Database::Track.count).to eq tracks.size
       end
 
+      context 'with series information' do
+        let(:s) { series[1] }
+
+        before do
+          db.add_series(s)
+        end
+
+        it 'adds a track with the series info' do
+          db.add({track: tracks[0], series: s})
+          def_title = tracks[0][:default_title]
+          t = Viva::Database::Track.where(default_title: def_title)
+          correct_series = Viva::Database::Series.where(raw: s[:raw]).first
+          expect(t.first.series).to eq correct_series
+        end
+      end
     end
 
-    context 'series only' do
-      it 'adds one series to the database at a time' do
+    context 'when adding series' do
+      it 'adds one' do
         db.add({series: series[0]})
-        db.add({series: series[0]})
-        expect(Viva::Database::Series.count).to eq 2
+        expect(Viva::Database::Series.count).to eq 1
       end
 
-      it 'adds multiple series to the database' do
+      it 'adds multiple' do
         db.add({series: series})
         expect(Viva::Database::Series.count).to eq series.count
       end
     end
 
-    context 'add track with series related' do
-      let(:s) { series[1] }
-
-      before do
-        db.add_series(s)
-      end
-
-      it 'adds a track with the series info' do
-        db.add({track: tracks[0], series: s})
-        t = Viva::Database::Track.where(default_title: tracks[0][:default_title])
-        correct_series = Viva::Database::Series.where(raw: s[:raw]).first
-        expect(t.first.series).to eq correct_series
-      end
-    end
-
-    context 'add nothing' do
-      it 'passes an empty Hash to the method' do
+    context 'when adding nothing' do
+      it 'adds nothing' do
         db.add({})
         expect(Viva::Database::Track.count).to eq 0
         expect(Viva::Database::Series.count).to eq 0
@@ -113,40 +111,37 @@ describe Viva::Database do
   end
 
   describe '#add_series' do
-    it 'should add one series to the database at a time' do
+    it 'adds one' do
       db.add_series(series[0])
-      db.add_series(series[0])
-      expect(Viva::Database::Series.count).to eq 2
+      expect(Viva::Database::Series.count).to eq 1
+    end
+
+    it 'adds multiple' do
+      db.add_series(series)
+      expect(Viva::Database::Series.count).to eq series.size
     end
   end
 
   describe '#add_tracks' do
     before(:each) do
+      Viva::Database::Track.delete_all unless Viva::Database::Track.count == 0
       db.add_series(series)
-      Viva::Database::Track.delete_all
     end
 
-    context 'one track at a time' do
-      it 'adds tracks to the database' do
-        db.add_tracks(tracks[0])
-        db.add_tracks(tracks[1])
-        db.add_tracks(tracks[2])
-
-        expect(Viva::Database::Track.count).to eq 3
-      end
+    it 'adds one' do
+      db.add_tracks(tracks[0])
+      expect(Viva::Database::Track.count).to eq 1
     end
 
-    context 'multiple tracks at once' do
-      it 'adds multiple tracks at once' do
-        db.add_tracks(tracks)
-        expect(Viva::Database::Track.count).to eq tracks.size
-      end
+    it 'adds multiple' do
+      db.add_tracks(tracks)
+      expect(Viva::Database::Track.count).to eq tracks.size
     end
 
-    context 'associate track with series' do
+    context 'with series information' do
       let (:dummy_series) { series[0][:raw] }
 
-      it 'associates multiple tracks with a series Hash ' do
+      it 'adds one ' do
         # Make the associated series for the tracks this
         db.add_tracks(tracks, {raw: dummy_series})
         # All of them should be associated with `dummy_series'
@@ -155,17 +150,13 @@ describe Viva::Database do
         expect(matches.count).to eq tracks.size
       end
 
-      it 'relates a track with a Series object' do
+      it 'adds multiple' do
         series = Viva::Database::Series.all.sample
         t = tracks.sample
         db.add_tracks(t, series)
         new_t = Viva::Database::Track.where(default_title: t[:default_title])
         expect(new_t.first.series).to eq series
       end
-    end
-
-    after(:each) do
-      Viva::Database::Track.delete_all
     end
   end
 
@@ -175,53 +166,60 @@ describe Viva::Database do
       db.add_tracks(tracks)
     end
 
-    it 'updates a track' do
-      new_title = 'updated title'
-      to_update = tracks[1]
-      db.update_track(to_update, {title: new_title})
-      result = Viva::Database::Track.where(title: new_title).first
-      expect(result.default_title).to eq to_update[:default_title]
+    context 'when updating a track' do
+      it 'updates a track' do
+        new_title = 'updated title'
+        to_update = tracks[1]
+        db.update_track(to_update, {title: new_title})
+        result = Viva::Database::Track.where(title: new_title).first
+        expect(result.default_title).to eq to_update[:default_title]
+      end
+
+      context 'with a series' do
+        context 'with a Series object' do
+        it do
+          to_update = Viva::Database::Track.all.sample
+          new_series = Viva::Database::Series.all.sample
+          db.update_track(to_update, {}, new_series)
+          updated = Viva::Database::Track.find(to_update.id)
+          expect(updated.series).to eq new_series
+        end
+        end
+
+        context 'with a Hash' do
+          it do
+            to_update = Viva::Database::Track.all.sample
+            new_series = series.sample
+            db.update_track(to_update, {}, {raw: new_series[:raw]})
+            updated = Viva::Database::Track.find(to_update.id)
+            expect(updated.series.jpn).to eq new_series[:jpn]
+          end
+        end
+      end
     end
 
-    it 'updates a series' do
-      new_title = 'アンパンマン、新しいタイトルよ！'
-      to_update = series.sample
-      db.update_series(to_update, {jpn: new_title})
-      result = Viva::Database::Series.where(jpn: new_title).first
-      expect(result.raw).to eq to_update[:raw]
+    context 'when updating a series' do
+      it 'updates a series' do
+        new_title = 'アンパンマン、新しいタイトルよ！'
+        to_update = series.sample
+        db.update_series(to_update, {jpn: new_title})
+        result = Viva::Database::Series.where(jpn: new_title).first
+        expect(result.raw).to eq to_update[:raw]
+      end
     end
 
-    it 'updates to an empty hash' do
-      to_update = Viva::Database::Track.all.sample
-      db.update_track(to_update, {})
-      updated = Viva::Database::Track.find(to_update.id)
-      expect(updated).to eq to_update
-    end
-
-    it 'updates a track related to a Series object' do
-      to_update = Viva::Database::Track.all.sample
-      new_series = Viva::Database::Series.all.sample
-      db.update_track(to_update, {}, new_series)
-      updated = Viva::Database::Track.find(to_update.id)
-      expect(updated.series).to eq new_series
-    end
-
-    it 'updates a track related to a series Hash' do
-      to_update = Viva::Database::Track.all.sample
-      new_series = series.sample
-      db.update_track(to_update, {}, {raw: new_series[:raw]})
-      updated = Viva::Database::Track.find(to_update.id)
-      expect(updated.series.jpn).to eq new_series[:jpn]
-    end
-
-    after(:each) do
-      Viva::Database::Series.delete_all
-      Viva::Database::Track.delete_all
+    context 'when updating nothing' do
+      it 'does nothing' do
+        to_update = Viva::Database::Track.all.sample
+        db.update_track(to_update, {})
+        updated = Viva::Database::Track.find(to_update.id)
+        expect(updated).to eq to_update
+      end
     end
   end
 
-  describe 'search' do
-    describe 'terms' do
+  context 'when searching' do
+    context 'with terms' do
       before do
         db.add_series(series)
 
@@ -257,7 +255,7 @@ describe Viva::Database do
       end
     end
 
-    describe 'strict search' do
+    context 'with strict search' do
       before do
         db.add_series(series)
         raw_series = series.map { |s| {raw: s[:raw]} }
@@ -267,64 +265,78 @@ describe Viva::Database do
       end
 
       describe '#search_strict' do
-        it 'searches using one condition' do
-          conditions = {series: {jpn: '高校'}}
-          result = db.search_strict(conditions)
-          # Just one because no tracks for 男子高校の日常 in seed data
-          expect(result.count).to eq 1
+        context 'with one condition' do
+          it do
+            conditions = {series: {jpn: '高校'}}
+            result = db.search_strict(conditions)
+            # Just one because no tracks for 男子高校の日常 in seed data
+            expect(result.count).to eq 1
+          end
         end
 
-        it 'searches using two conditions' do
-          conditions = {
-            series: {
-              jpn: '高校',
-              eng: 'Magic',
+        context 'with two conditions' do
+          it do
+            conditions = {
+              series: {
+                jpn: '高校',
+                eng: 'Magic',
+              }
             }
-          }
-          result = db.search_strict(conditions)
-          expect(result.first.series.raw).to eq 'mahouka-koukou-no-rettousei'
+            result = db.search_strict(conditions)
+            expect(result.first.series.raw).to eq 'mahouka-koukou-no-rettousei'
+          end
         end
 
-        it 'searches with both series and track' do
-          conditions = {
-            series: {
-              jpn: '高校',
-            },
-            track: {
-              default_title: 'hope',
-            },
-          }
-          result = db.search_strict(conditions)
-          expect(result.first.title).to eq 'Rising Hope'
+        context 'with both series and track' do
+          it do
+            conditions = {
+              series: {
+                jpn: '高校',
+              },
+              track: {
+                default_title: 'hope',
+              },
+            }
+            result = db.search_strict(conditions)
+            expect(result.first.title).to eq 'Rising Hope'
+          end
         end
       end
 
       describe '#search_series_strict' do
-        it 'searches for a series using one condition' do
-          conditions = {jpn: '高校'}
-          result = db.search_series_strict(conditions)
-          expect(result.count).to eq 2
+        context 'with one condition' do
+          it do
+            conditions = {jpn: '高校'}
+            result = db.search_series_strict(conditions)
+            expect(result.count).to eq 2
+          end
         end
 
-        it 'searches for a series using two conditions' do
-          conditions = {jpn: '高校', eng: 'Magic'}
-          result = db.search_series_strict(conditions)
-          expect(result.first.raw).to eq 'mahouka-koukou-no-rettousei'
+        context 'with two conditions' do
+          it do
+            conditions = {jpn: '高校', eng: 'Magic'}
+            result = db.search_series_strict(conditions)
+            expect(result.first.raw).to eq 'mahouka-koukou-no-rettousei'
+          end
         end
       end
 
       describe '#search_tracks_strict' do
-        it 'searches for a track using one condition' do
-          conditions = {default_title: 'rising'}
-          results = db.search_tracks_strict(conditions)
-          expect(results.first.artist).to eq 'LiSA'
+        context 'with one condition' do
+          it do
+            conditions = {default_title: 'rising'}
+            results = db.search_tracks_strict(conditions)
+            expect(results.first.artist).to eq 'LiSA'
+          end
         end
 
-        it 'searches for a track using two conditions' do
-          conditions = {default_title: 'hope', url: '164'}
-          results = db.search_tracks_strict(conditions)
-          expected_title = 'resuscitated hope -tv size-'
-          expect(results.first.default_title).to eq expected_title
+        context 'with two conditions' do
+          it do
+            conditions = {default_title: 'hope', url: '164'}
+            results = db.search_tracks_strict(conditions)
+            expected_title = 'resuscitated hope -tv size-'
+            expect(results.first.default_title).to eq expected_title
+          end
         end
       end
     end
@@ -337,20 +349,24 @@ describe Viva::Database do
       db.add_tracks(tracks[0], raw_series[0])
     end
 
-    it 'returns a simple string representation' do
-      track = Viva::Database::Track.where(default_title: 'rising hope')
-      str = track.first.to_s
-      expect(str).to eq 'Rising Hope from 魔法科高校の劣等生'
+    context 'without detail' do
+      it 'returns a simple string representation' do
+        track = Viva::Database::Track.where(default_title: 'rising hope')
+        str = track.first.to_s
+        expect(str).to eq 'Rising Hope from 魔法科高校の劣等生'
+      end
     end
 
-    it 'returns a detailed string representation' do
-      track = Viva::Database::Track.where(default_title: 'rising hope')
-      str = track.first.to_s(detailed: true)
-      expected = <<EOF
+    context 'with detail' do
+      it 'returns a detailed string representation' do
+        track = Viva::Database::Track.where(default_title: 'rising hope')
+        str = track.first.to_s(detailed: true)
+        expected = <<EOF
 Rising Hope from 魔法科高校の劣等生
 Artist: LiSA
 EOF
-      expect(str).to eq expected.strip
+        expect(str).to eq expected.strip
+      end
     end
   end
